Change instantiation code style
If no other identifier is specifying an instantion of a module, it
should be used "<entity_name>_<n>", where n is the number, which
identify each of individual existing instantions of the same module. It
starts from 0.

To declare signals associated to the instantioned module in the
implementing one, they name must have the
"<instantion_name_shortcut>_<port_associated_name>" format. The "<n>" of
the instation name is used only when it is required to instantiate more
than one of the instations.

See seg7_driver.vhd:46-48 and seg7_driver.vhd:56-61 for example.