Update the FIFO module
The FIFO's behavior has been fixed, the problem was with being FIFO in
the full state while write and read at the same time, the full state
did not persist.

The file_tb.vhd has been updated to test the issue above and it's code
has been edited to be more clear.