Update code style
The update has affected the following:
1. The file with implementation now initialize the generic data with
values equivalent with it's simulation file data to be synthesizable.
2. Some port data types have been changed to clear their meaning as
std_logic_vector can mean anything and often it is not so complicated to
cast/convert one data type to another.
3. Mostly some constant redundancy dependencies have been removed.
Mostly it means replace <A_CONSTANT> with attribute of a data object, if
possible.