Update code style
The prevention process in implemented modules has been divided into
input_prevention, which prevent input values leading to undefined
internal state, and output_prevention, which checks output of the module
with using local buffers.

Also the report string contains a space on the beginning of the next
line, if any.