Initial ROM initialization from a file
The first version of the ROM initialization from a file has been
implemented. It has been based on the previous rom's module interface,
mostly on g_INIT_VECTOR.

Unfortunately, it does not seems like an optimal solution and it will be
remaked soon in a new branch together with the rom's interface and
implementation, which will be merged to the master then.

It is highly possible that the util and verif_util packages will be
remastered as well.