Use relative compare with now in ram module
This also resulted in creating a new file - ram_tb_public.vhd. It
contains a package, which holds the c_CLK_PERIOD constant. This
constant needs to be shared with ram module, because the need of the
result of compare with now function to be relative to c_CLK_PERIOD.